Last commit missed tracker changes
Also altered the tracker so it will now accept and appropriately handle
NodeRequest messages with a list of already-known IDs (so it knows to
send a new one). Also, this should go without saying but the P2PNode
changes completely broke all the unit tests, which will need to be
rewritten. Cheers!